LOL we didn't fix the threads logging issue
There was a duplicate message logging because each of these aux methods
calls log_request and the `/data` endpoint itself was also calling it.

There are more log messages than beforenow , but the test failure was
correct. Threads still don't get their log output sent to stdout. Need
to flush something? IDK, it's not important to the function of the app.

Time to go outside!

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>