Miscellaneous fixes
* sleep 1 -> sleep 0 (I think this will still work)

I read in some documentation that "sleep 0" is a signal to the GVL that
another thread waiting should be allowed to take control, here's a link:

https://stackoverflow.com/questions/36753094/sleep-0-has-special-meaning

* cron_interval should be set to CACHE_DURATION

We set it, as it is set in the tests, to wake up the cron thread as
needed. With cron_interval set at 0, the cron thread never wakes at all!

* Indentation errors and whitespace

The cursor editor has been reconfigured to expect/preserve newlines at
the end of the file, so we should not have editors to fight with each
other anymore as I switch between cursor and vim (phew!)

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>